+++ src/utils/indent.ts
@@ -1,7 +1,4 @@
-export const indent = (str: string, spaces: number) => {
-  let indentation = '';
-  for (let i = 0; i < spaces; i++) {
-    indentation += ' ';
-  }
-  return str.replace(/^/mg, indentation);
-};
+import {repeat} from './repeat';
+
+export const indent = (str: string, spaces: number) =>
+  str.replace(/^/mg, repeat(' ', spaces));